Convert existing project to gradle project
