+++ .gitignore
@@ -1,6 +1,8 @@
 # Covers JetBrains IDEs: IntelliJ, RubyMine, PhpStorm, AppCode, PyCharm, CLion, Android Studio, WebStorm and Rider
 # Reference: https://intellij-support.jetbrains.com/hc/en-us/articles/206544839
 
+.gradle
+
 # User-specific stuff
 .idea/**/workspace.xml
 .idea/**/tasks.xml
