+++ src/main/java/org/iogame/main/Main.java
@@ -1,6 +1,4 @@
-package iogame.main;
-
-import iogame.main.Game;
+package org.iogame.main;
 
 public class Main {
