+++ src/main/java/org/iogame/model/enums/EResource.java
@@ -1,4 +1,4 @@
-package iogame.model.enums;
+package org.iogame.model.enums;
 
 public enum EResource {
     RES1, RES2, RES3
