+++ src/main/java/org/iogame/model/planet/Planet.java
@@ -1,4 +1,4 @@
-package iogame.model.planet;
+package org.iogame.model.planet;
 
 public class Planet {
     private double x;
