+++ src/main/java/org/iogame/model/planet/buildings/Building.java
@@ -1,6 +1,7 @@
-package iogame.model.planet.buildings;
+package org.iogame.model.planet.buildings;
 
-import iogame.model.enums.EResource;
+
+import org.iogame.model.enums.EResource;
 
 import java.util.HashMap;
